+++ src/documentation.ts
@@ -18,6 +18,7 @@ import { CommonConfig } from "./config-type";
 import { mapLogicalContainer } from "./logical-container";
 import { Method } from "./method";
 import {
+  OpenAPIContext,
   depictRequest,
   depictRequestParams,
   depictResponse,
@@ -29,6 +30,7 @@ import {
 } from "./documentation-helpers";
 import { Routing } from "./routing";
 import { RoutingWalkerParams, walkRouting } from "./routing-walker";
+import { HandlingRules } from "./schema-walker";
 
 type Component =
   | "positiveResponse"
@@ -64,6 +66,13 @@ interface DocumentationParams {
    * @default JSON.stringify() + SHA1 hash as a hex digest
    * */
   serializer?: (schema: z.ZodTypeAny) => string;
+  /**
+   * @desc Handling rules for your own branded schemas.
+   * @desc Keys: brands (recommended to use unique symbols).
+   * @desc Values: functions having schema as first argument that you should assign type to, second one is a context.
+   * @example { MyBrand: ( schema: typeof myBrandSchema, { next } ) => ({ type: "object" })
+   */
+  brandHandling?: HandlingRules<SchemaObject | ReferenceObject, OpenAPIContext>;
 }
 
 export class Documentation extends OpenApiBuilder {
@@ -132,6 +141,7 @@ export class Documentation extends OpenApiBuilder {
     version,
     serverUrl,
     descriptions,
+    brandHandling,
     hasSummaryFromDescription = true,
     composition = "inline",
     serializer = defaultSerializer,
@@ -153,6 +163,7 @@ export class Documentation extends OpenApiBuilder {
         endpoint,
         composition,
         serializer,
+        brandHandling,
         getRef: this.getRef.bind(this),
         makeRef: this.makeRef.bind(this),
       };