+++ src/index.ts
@@ -35,6 +35,10 @@ export { Integration } from "./integration";
 
 export { ez } from "./proprietary-schemas";
 
+// Convenience types
+export type { Depicter } from "./documentation-helpers";
+export type { Producer } from "./zts-helpers";
+
 // Issues 952, 1182, 1269: Insufficient exports for consumer's declaration
 export type { MockOverrides } from "./testing";
 export type { Routing } from "./routing";