+++ src/zts.ts
@@ -4,6 +4,7 @@ import { hasCoercion, tryToTransform } from "./common-helpers";
 import { ezDateInBrand } from "./date-in-schema";
 import { ezDateOutBrand } from "./date-out-schema";
 import { FileSchema, ezFileBrand } from "./file-schema";
+import { ProprietaryBrand } from "./proprietary-schemas";
 import { RawSchema, ezRawBrand } from "./raw-schema";
 import { HandlingRules, walkSchema } from "./schema-walker";
 import {
@@ -26,9 +27,7 @@ const samples = {
   [ts.SyntaxKind.UndefinedKeyword]: undefined,
 } satisfies Partial<Record<ts.KeywordTypeSyntaxKind, unknown>>;
 
-const onLiteral: Producer<z.ZodLiteral<LiteralType>> = ({
-  schema: { value },
-}) =>
+const onLiteral: Producer = ({ value }: z.ZodLiteral<LiteralType>) =>
   f.createLiteralTypeNode(
     typeof value === "number"
       ? f.createNumericLiteral(value)
@@ -39,12 +38,14 @@ const onLiteral: Producer<z.ZodLiteral<LiteralType>> = ({
         : f.createStringLiteral(value),
   );
 
-const onObject: Producer<z.ZodObject<z.ZodRawShape>> = ({
-  schema: { shape },
-  isResponse,
-  next,
-  optionalPropStyle: { withQuestionMark: hasQuestionMark },
-}) => {
+const onObject: Producer = (
+  { shape }: z.ZodObject<z.ZodRawShape>,
+  {
+    isResponse,
+    next,
+    optionalPropStyle: { withQuestionMark: hasQuestionMark },
+  },
+) => {
   const members = Object.entries(shape).map<ts.TypeElement>(([key, value]) => {
     const isOptional =
       isResponse && hasCoercion(value)
@@ -66,33 +67,32 @@ const onObject: Producer<z.ZodObject<z.ZodRawShape>> = ({
   return f.createTypeLiteralNode(members);
 };
 
-const onArray: Producer<z.ZodArray<z.ZodTypeAny>> = ({
-  schema: { element },
-  next,
-}) => f.createArrayTypeNode(next(element));
+const onArray: Producer = ({ element }: z.ZodArray<z.ZodTypeAny>, { next }) =>
+  f.createArrayTypeNode(next(element));
 
-const onEnum: Producer<z.ZodEnum<[string, ...string[]]>> = ({
-  schema: { options },
-}) =>
+const onEnum: Producer = ({ options }: z.ZodEnum<[string, ...string[]]>) =>
   f.createUnionTypeNode(
     options.map((option) =>
       f.createLiteralTypeNode(f.createStringLiteral(option)),
     ),
   );
 
-const onSomeUnion: Producer<
-  | z.ZodUnion<z.ZodUnionOptions>
-  | z.ZodDiscriminatedUnion<string, z.ZodDiscriminatedUnionOption<string>[]>
-> = ({ schema: { options }, next }) => f.createUnionTypeNode(options.map(next));
+const onSomeUnion: Producer = (
+  {
+    options,
+  }:
+    | z.ZodUnion<z.ZodUnionOptions>
+    | z.ZodDiscriminatedUnion<string, z.ZodDiscriminatedUnionOption<string>[]>,
+  { next },
+) => f.createUnionTypeNode(options.map(next));
 
 const makeSample = (produced: ts.TypeNode) =>
   samples?.[produced.kind as keyof typeof samples];
 
-const onEffects: Producer<z.ZodEffects<z.ZodTypeAny>> = ({
-  schema,
-  next,
-  isResponse,
-}) => {
+const onEffects: Producer = (
+  schema: z.ZodEffects<z.ZodTypeAny>,
+  { next, isResponse },
+) => {
   const input = next(schema.innerType());
   const effect = schema._def.effect;
   if (isResponse && effect.type === "transform") {
@@ -114,7 +114,7 @@ const onEffects: Producer<z.ZodEffects<z.ZodTypeAny>> = ({
   return input;
 };
 
-const onNativeEnum: Producer<z.ZodNativeEnum<z.EnumLike>> = ({ schema }) =>
+const onNativeEnum: Producer = (schema: z.ZodNativeEnum<z.EnumLike>) =>
   f.createUnionTypeNode(
     Object.values(schema.enum).map((value) =>
       f.createLiteralTypeNode(
@@ -125,11 +125,10 @@ const onNativeEnum: Producer<z.ZodNativeEnum<z.EnumLike>> = ({ schema }) =>
     ),
   );
 
-const onOptional: Producer<z.ZodOptional<z.ZodTypeAny>> = ({
-  next,
-  schema,
-  optionalPropStyle: { withUndefined: hasUndefined },
-}) => {
+const onOptional: Producer = (
+  schema: z.ZodOptional<z.ZodTypeAny>,
+  { next, optionalPropStyle: { withUndefined: hasUndefined } },
+) => {
   const actualTypeNode = next(schema.unwrap());
   return hasUndefined
     ? f.createUnionTypeNode([
@@ -139,84 +138,74 @@ const onOptional: Producer<z.ZodOptional<z.ZodTypeAny>> = ({
     : actualTypeNode;
 };
 
-const onNullable: Producer<z.ZodNullable<z.ZodTypeAny>> = ({ next, schema }) =>
+const onNullable: Producer = (schema: z.ZodNullable<z.ZodTypeAny>, { next }) =>
   f.createUnionTypeNode([
     next(schema.unwrap()),
     f.createLiteralTypeNode(f.createNull()),
   ]);
 
-const onTuple: Producer<z.ZodTuple> = ({
-  next,
-  schema: {
-    items,
-    _def: { rest },
-  },
-}) =>
+const onTuple: Producer = ({ items, _def: { rest } }: z.ZodTuple, { next }) =>
   f.createTupleTypeNode(
     items
       .map(next)
       .concat(rest === null ? [] : f.createRestTypeNode(next(rest))),
   );
 
-const onRecord: Producer<z.ZodRecord> = ({
-  next,
-  schema: { keySchema, valueSchema },
-}) =>
+const onRecord: Producer = (
+  { keySchema, valueSchema }: z.ZodRecord,
+  { next },
+) =>
   f.createExpressionWithTypeArguments(
     f.createIdentifier("Record"),
     [keySchema, valueSchema].map(next),
   );
 
-const onIntersection: Producer<
-  z.ZodIntersection<z.ZodTypeAny, z.ZodTypeAny>
-> = ({ next, schema }) =>
-  f.createIntersectionTypeNode([schema._def.left, schema._def.right].map(next));
+const onIntersection: Producer = (
+  { _def }: z.ZodIntersection<z.ZodTypeAny, z.ZodTypeAny>,
+  { next },
+) => f.createIntersectionTypeNode([_def.left, _def.right].map(next));
 
-const onDefault: Producer<z.ZodDefault<z.ZodTypeAny>> = ({ next, schema }) =>
-  next(schema._def.innerType);
+const onDefault: Producer = ({ _def }: z.ZodDefault<z.ZodTypeAny>, { next }) =>
+  next(_def.innerType);
 
 const onPrimitive =
-  (syntaxKind: ts.KeywordTypeSyntaxKind): Producer<z.ZodTypeAny> =>
+  (syntaxKind: ts.KeywordTypeSyntaxKind): Producer =>
   () =>
     f.createKeywordTypeNode(syntaxKind);
 
-const onBranded: Producer<
-  z.ZodBranded<z.ZodTypeAny, string | number | symbol>
-> = ({ next, schema }) => next(schema.unwrap());
+const onBranded: Producer = (
+  schema: z.ZodBranded<z.ZodTypeAny, string | number | symbol>,
+  { next },
+) => next(schema.unwrap());
 
-const onReadonly: Producer<z.ZodReadonly<z.ZodTypeAny>> = ({ next, schema }) =>
-  next(schema._def.innerType);
+const onReadonly: Producer = (schema: z.ZodReadonly<z.ZodTypeAny>, { next }) =>
+  next(schema.unwrap());
 
-const onCatch: Producer<z.ZodCatch<z.ZodTypeAny>> = ({ next, schema }) =>
-  next(schema._def.innerType);
+const onCatch: Producer = ({ _def }: z.ZodCatch<z.ZodTypeAny>, { next }) =>
+  next(_def.innerType);
 
-const onPipeline: Producer<z.ZodPipeline<z.ZodTypeAny, z.ZodTypeAny>> = ({
-  schema,
-  next,
-  isResponse,
-}) => next(schema._def[isResponse ? "out" : "in"]);
+const onPipeline: Producer = (
+  { _def }: z.ZodPipeline<z.ZodTypeAny, z.ZodTypeAny>,
+  { next, isResponse },
+) => next(_def[isResponse ? "out" : "in"]);
 
-const onNull: Producer<z.ZodNull> = () =>
-  f.createLiteralTypeNode(f.createNull());
+const onNull: Producer = () => f.createLiteralTypeNode(f.createNull());
 
-const onLazy: Producer<z.ZodLazy<z.ZodTypeAny>> = ({
-  getAlias,
-  makeAlias,
-  next,
-  serializer: serialize,
-  schema: lazy,
-}) => {
-  const name = `Type${serialize(lazy.schema)}`;
+const onLazy: Producer = (
+  { schema }: z.ZodLazy<z.ZodTypeAny>,
+  { getAlias, makeAlias, next, serializer: serialize },
+) => {
+  const name = `Type${serialize(schema)}`;
   return (
     getAlias(name) ||
     (() => {
       makeAlias(name, f.createLiteralTypeNode(f.createNull())); // make empty type first
-      return makeAlias(name, next(lazy.schema)); // update
+      return makeAlias(name, next(schema)); // update
     })()
   );
 };
 
-const onFile: Producer<FileSchema> = ({ schema }) => {
+const onFile: Producer = (schema: FileSchema) => {
   const subject = schema.unwrap();
   const stringType = f.createKeywordTypeNode(ts.SyntaxKind.StringKeyword);
   const bufferType = f.createTypeReferenceNode("Buffer");
@@ -228,10 +217,14 @@ const onFile: Producer<FileSchema> = ({ schema }) => {
       : bufferType;
 };
 
-const onRaw: Producer<RawSchema> = ({ next, schema }) =>
+const onRaw: Producer = (schema: RawSchema, { next }) =>
   next(schema.unwrap().shape.raw);
 
-const producers: HandlingRules<ts.TypeNode, ZTSContext> = {
+const producers: HandlingRules<
+  ts.TypeNode,
+  ZTSContext,
+  z.ZodFirstPartyTypeKind | ProprietaryBrand
+> = {
   ZodString: onPrimitive(ts.SyntaxKind.StringKeyword),
   ZodNumber: onPrimitive(ts.SyntaxKind.NumberKeyword),
   ZodBigInt: onPrimitive(ts.SyntaxKind.BigIntKeyword),
@@ -263,15 +256,18 @@ const producers: HandlingRules<ts.TypeNode, ZTSContext> = {
   [ezRawBrand]: onRaw,
 };
 
-export const zodToTs = ({
-  schema,
-  ...ctx
-}: {
-  schema: z.ZodTypeAny;
-} & ZTSContext) =>
-  walkSchema<ts.TypeNode, ZTSContext>({
-    schema,
-    rules: producers,
+export const zodToTs = (
+  schema: z.ZodTypeAny,
+  {
+    brandHandling,
+    ctx,
+  }: {
+    brandHandling?: HandlingRules<ts.TypeNode, ZTSContext>;
+    ctx: ZTSContext;
+  },
+) =>
+  walkSchema(schema, {
+    rules: { ...brandHandling, ...producers },
     onMissing: () => f.createKeywordTypeNode(ts.SyntaxKind.AnyKeyword),
-    ...ctx,
+    ctx,
   });